Ajout des tokens JWT et des rôles via Identity

+++ Findexium/DTOs/BidListDTO.cs
@@ -8,10 +8,10 @@ namespace Findexium.DTOs
         public int BidListId { get; set; }
         public string? Account { get; set; }
         public string? BidType { get; set; }
-        public double? BidQuantity { get; set; }
-        public double? AskQuantity { get; set; }
-        public double? Bid { get; set; }
-        public double? Ask { get; set; }
+        public decimal? BidQuantity { get; set; }
+        public decimal? AskQuantity { get; set; }
+        public decimal? Bid { get; set; }
+        public decimal? Ask { get; set; }
         public string? Benchmark { get; set; }
         public DateTime? BidListDate { get; set; }
         public string? Commentary { get; set; }
