+++ Findexium/Domain/Trade.cs
@@ -5,10 +5,10 @@ namespace Findexium.Domain
         public int TradeId { get; set; }
         public string? Account { get; set; }
         public string? AccountType { get; set; }
-        public double? BuyQuantity { get; set; }
-        public double? SellQuantity { get; set; }
-        public double? BuyPrice { get; set; }
-        public double? SellPrice { get; set; }
+        public decimal? BuyQuantity { get; set; }
+        public decimal? SellQuantity { get; set; }
+        public decimal? BuyPrice { get; set; }
+        public decimal? SellPrice { get; set; }
         public DateTime? TradeDate { get; set; }
         public string? TradeSecurity { get; set; }
         public string? TradeStatus { get; set; }
